Moves hover index handling inside AreaGraph
